Untangles the Rune Sword a little bit.
Instead of having different damage and attack speed whether in the different modes, this just makes it so the only thing toggling the sword does is allow you to shoot the ranged bit. 

This'll solve problems like trying to melee stuff and firing the projectile and not being able to attack with that hand for a few seconds.