Adds some special snowflake masks for special snowflakes. (#333)
* adds scuffed sprites

scuffed some of the sprites

but honestly from up close it looks scuffed in general

* adds some scuffed code

I am aware of the bugs it may have.

hmm today I will

* updates the mask a bit

i hate the 14th sprite

* removes defense value

* adds some comments on my thought process

* oopsie whoopsies

I forgot to set weight to 0! How silly of me!
Al;though this might not be reflected in game.