Courier Coats (#417)
* Courier Coats

Adds coats made by Golgor#7400

they share the stats of the Burger coat, and activates the Burger Coat code

* ignore the holy code in the DME... its an error on my end with swapping branches

reeee

* adds knuckles where they go

adds them to the cultist and ancient cultist for the dark knuckles, and the chap vendor for the holy ones

* Adds Old Glory

adds in Old Glory as a polearm style melee weapon like the Goendendag with more scaling off of melee and strength so its still viable late game.... well compared to the Goendendag at least. Also adds it to the vendor alongside all the coats.