adds energy wrench and gold toolbox
Adds in a special energy wrench that is basically a reskinned esword. once we get things we can use wrenches on, i will try and fix the flag code for it.

Also adds in the LEGENDARY GOLDEN TOOLBOX! its fairly chunky as melee clubs go, and can only be found in the trash! Good hunting!