changes some three icons. Read before merge.
What does it do?: Changes icons.

Note: The new knuckles icon is already two knuckles in the icon: Therefore I recommend you just make it how you can use fire axes by just two-handing them. The previous two sprites are still in the sprite file, and have not be removed.

It can be fun for the martial-artists in us to use knuckles, but annoying to manage because of the small sprite and generally having to carry around two: Therefore I just picked one.

P.S. I don't know why but the axe folder seemingly wasn't there. I don't know if that's an issue to be corrected or not.